Add notifications for incoming calls in background
A way to open the WebSocket when a call is received when the app is in the
background is required. Also, the web socket myst be closed when a call that was
received in the background is declined.
Added two notifications for that, which trigger the ApplicationLifecycleObserver
to open and close the web socket. It is the implementer's responsibility to post
the notifications at the appropriate time when a background call occurs.